pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "options-agent"
version = "0.2.0"
description = "AI-powered options trading analysis agent"
requires-python = ">=3.11"
dependencies = [
    "polygon-api-client",
    "alpaca-py",
    "langchain",
    "langgraph",
    "langchain-anthropic",
    "pandas",
    "numpy",
    "ta",
    "python-dotenv",
    "rich",
    "fastapi",
    "uvicorn",
    "schedule",
    "pydantic",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
include = ["tools*", "agents*", "api*", "jobs*", "data*"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
